backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-multipart
passlib[bcrypt]
PyJWT
alembic
httpx
pytest
